requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo==4.6.3
transformers==4.39.3
torch==2.2.2
numpy<2.0
python-dotenv==1.0.1
